+++ PIconnect/test/__init__.py
@@ -1,7 +1,0 @@
-""" PIconnect.test
-    Main module for the PIconnect.test package.
-"""
-import unittest
-
-if __name__ == '__main__':
-    unittest.main()