+++ HISTORY.rst
@@ -2,6 +2,25 @@
 History
 =======
 
+0.7.1 (?)
+---------
+
+* Improved documentation
+* Changed `PIData.PISeriesContainer` to an Abstract Base Class
+
+0.7.0 (2018-11-14)
+------------------
+
+* Add `summary`, `summaries`, and `filtered_summaries` methods to `PIPoint`
+    and `PIAFAttribute`
+
+0.6.0 (2018-07-05)
+------------------
+
+0.5.1 (2017-11-25)
+------------------
+
+
 0.4.0 (2017-11-25)
 ------------------